winpthreads: Prevent scoping issues and warnings in cleanup_push/pop
Interestingly, pthread_cleanup_push/pop are allowed by POSIX to be
implemented as macros opening and closing an lexical scope. By using
the well-known trick of do { ... } while (0) [1], we prevent potential
scoping issues in surrounding code when the macro is expanded.

Removing the comma operator in pthread_cleanup_pop prevents compiler
warnings against not using the return value of the expression, and a
against an empty statement.

[1]: https://gcc.gnu.org/onlinedocs/cpp/Swallowing-the-Semicolon.html

Signed-off-by: Antonin Décimo <antonin@tarides.com>
Signed-off-by: Martin Storsjö <martin@martin.st>